refactor(apiResponseMapper): These utilities were included
