+++ authorization/src/main/java/com/jobby/authorization/infraestructure/adapters/in/rest/AuthorizationController.java
@@ -4,7 +4,7 @@ import com.jobby.authorization.domain.ports.in.AuthorizeEmployeeByCredentials;
 import com.jobby.authorization.domain.ports.in.AuthorizeEmployeeByTokens;
 import com.jobby.authorization.domain.ports.out.SafeResultValidator;
 import com.jobby.domain.mobility.validator.ValidationChain;
-import com.jobby.authorization.infraestructure.response.definition.ApiResponseMapper;
+import com.jobby.infraestructure.response.definition.ApiResponseMapper;
 import com.jobby.authorization.infraestructure.dto.mappers.TokenRegistryResponseMapper;
 import com.jobby.authorization.infraestructure.dto.requests.LoginRequest;
 import com.jobby.authorization.infraestructure.dto.requests.TokenRequest;
